Add files via upload
It's better to use Kansa with JSON as an output format to make sure this module yields the result. This has to do with the way the data is sent from remote data pipeline and then processed by powershell. If you run with CSV as output, you won't be able to received complete data.